# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Cuadra API
  version: "1.0.0"
  description: |
    REST API del backend de Cuadra POS.
    Multi-tenant: todos los endpoints filtran por `ownerId` del JWT.
    Auth: Bearer JWT en header `Authorization`.
servers:
  - url: http://localhost:4000
    description: Desarrollo local
  - url: https://cuadra-backend.railway.app
    description: Producción (Railway)

# ─── SECURITY ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SaleItemCreate:
      type: object
      required: [product_id, product_name, quantity, price, final_price]
      properties:
        product_id: { type: string }
        product_name: { type: string }
        quantity: { type: integer, minimum: 1 }
        price: { type: number }
        cost_price: { type: number, nullable: true }
        variant_id: { type: string, nullable: true }
        variant_name: { type: string, nullable: true }
        final_price: { type: number }
        discount_applied: { type: string, nullable: true }
    SaleItemResponse:
      allOf:
        - $ref: "#/components/schemas/SaleItemCreate"
        - type: object
          properties:
            id: { type: string }
            sale_id: { type: string }
    SaleCreate:
      type: object
      required: [items, total, payment_method]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SaleItemCreate" }
        total: { type: number }
        payment_method:
          type: string
          enum: [cash, transfer, mobile_pay, credit]
        client_id: { type: string, nullable: true }
        client_name: { type: string, nullable: true }
        cashbox_id: { type: string, nullable: true }
        cashbox_name: { type: string, nullable: true }
        location_id: { type: string, nullable: true }
        location_name: { type: string, nullable: true }
        exchange_rate_at_sale: { type: number, nullable: true }
        notes: { type: string, nullable: true }
        discount_amount: { type: number, nullable: true }
        discount_reason: { type: string, nullable: true }
        promotion_savings: { type: number, nullable: true }
        coupon_code: { type: string, nullable: true }
        evidence_url: { type: string, nullable: true }
        payment_data:
          type: object
          nullable: true
          properties:
            reference: { type: string }
            bank: { type: string }
            date: { type: string }
    SaleResponse:
      allOf:
        - $ref: "#/components/schemas/SaleCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            status: { type: string, enum: [paid, pending, cancelled] }
            created_by: { type: string, nullable: true }
            creator_name: { type: string, nullable: true }
            created_at: { type: string, format: date-time }
            paid_at: { type: string, format: date-time, nullable: true }
            commission_pct: { type: number, nullable: true }
            commission_amount: { type: number, nullable: true }
            items:
              type: array
              items: { $ref: "#/components/schemas/SaleItemResponse" }
    CancelSaleRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    PayDebtsRequest:
      type: object
      required: [payment_method]
      properties:
        payment_method: { type: string, enum: [cash, transfer, mobile_pay] }
        cashbox_id: { type: string, nullable: true }

    # ── Clients ─────────────────────────────────────────────────────────────
...
    SupplierCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contact_name: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        address: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    SupplierResponse:
      allOf:
        - $ref: "#/components/schemas/SupplierCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            active: { type: boolean }

    # ── Categories ───────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []

# ─── PATHS ────────────────────────────────────────────────────────────────────
paths:
...
  /sales:
    get:
      summary: Listar ventas
      tags: [sales]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [paid, pending, cancelled] }
        - name: client_id
          in: query
          schema: { type: string }
        - name: start_date
          in: query
          schema: { type: string, format: date-time }
        - name: end_date
          in: query
          schema: { type: string, format: date-time }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SaleResponse" }
          description: OK
    post:
      summary: Crear venta (descuenta stock atómicamente)
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaleCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaleResponse" }
          description: Venta creada
        "400":
          description: Stock insuficiente u otros errores de negocio
  /sales/{saleId}:
    parameters:
      - name: saleId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Obtener venta por ID
      tags: [sales]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaleResponse" }
          description: OK
  /sales/{saleId}/status:
    patch:
      summary: Actualizar estado de venta (ej. pending → paid)
      tags: [sales]
      parameters:
        - name: saleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [paid, cancelled] }
                payment_method: { type: string }
                cashbox_id: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaleResponse" }
          description: OK
  /sales/{saleId}/cancel:
    post:
      summary: Cancelar venta (restaura stock)
      tags: [sales]
      parameters:
        - name: saleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CancelSaleRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SaleResponse" }
          description: OK
  /sales/clients/{clientId}/pay-all-debts:
    post:
      summary: Pagar todas las deudas de un cliente
      tags: [sales]
      parameters:
        - name: clientId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PayDebtsRequest" }
      responses:
        "200":
          description: Deudas saldadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: integer }

  # ── Clients ─────────────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      summary: Listar proveedores
      tags: [suppliers]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupplierResponse" }
          description: OK
    post:
      summary: Crear proveedor
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierResponse" }
          description: Proveedor creado
  /suppliers/{supplierId}:
    parameters:
      - name: supplierId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Actualizar proveedor
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierResponse" }
          description: OK
    delete:
      summary: Desactivar proveedor
      tags: [suppliers]
      responses:
        "204":
          description: Desactivado

  # ── Categories ────────────────────────────────────────────────────────────────
...